BPE26-ref: extract color utils and track quantifiers features to dedicated file
